Refactor core application to fix circular import and improve API key handling.
I've addressed two critical issues in the codebase:

1.  **Resolved Circular Dependency:**
    - I moved the `ProcessingJob` dataclass from the `orchestrator` module to a new, dedicated `types.py` file.
    - This resolves a circular import loop between the `orchestrator` and `ingestor` modules, improving the application's architecture and stability.

2.  **Improved API Key Handling:**
    - The application no longer relies on a hardcoded, uncommitted `api.py` file, which caused it to crash on startup.
    - The configuration now securely loads the `GEMINI_API_KEY` by first checking for an environment variable and then falling back to the `pkm_gardener/api.py` file.
    - I added a clear error message if the key is not found in either location.
    - I updated `.gitignore` to correctly ignore the local API key file.
    - I removed the obsolete `move_api_file.py` script.
