--- move_api_file.py
@@ -1,5 +1,0 @@
-import shutil
-import os
-
-shutil.move("D:\\PKM\\api.py", "D:\\PKM\\pkm_gardener\\api.py")
-
