+++ pkm_gardener/core_modules/indexer.py
@@ -1,4 +1,4 @@
-from pkm_gardener.orchestrator import ProcessingJob
+from pkm_gardener.types import ProcessingJob
 import os
 
 def update_index(job: ProcessingJob):
